fix(OwnerReferenceUpdates): Handle error logging for owner reference updates
Check for conflicts before logging an error for failed updates of owner
reference to the k8s secrets. This is necessary when multiple instances
of osm-controller are deployed, as they will all try to update the same
resource and only one of them will update it, which is sufficient.

Fixes #3776

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>