chore(security/scan): Scan docker images
Add step in the CI and the release job to
scan docker images for potential security
vulnerabilities. The Jobs will show all security
scan outputs in the logs, but will only exit
if the severity level is Medium, High, or Critical,
and will continue if if the level is Unknown or Low.

Signed-off-by: nshankar13 <nshankar@microsoft.com>