fix(grafana): use osm_request_duration_ms for latency graphs
This PR replaces envoy_cluster_upstream_rq_time with the existing
osm_request_duration_ms SMI metric to display latency in the mesh.
Currently, the latency graphs are present in the pod to service,
service to service, and workload to service dashboards. Unlike
envoy_cluster_upstream_rq_time, the osm_request_duration_ms metric
does not capture the src or destination service. Therefore, the
latency graphs no longer fit on the dashboards that allow the user
to specify a source service or see the latencies labeled with the
envoy cluster name (which includes the destination service name).

This PR removes the latency graphs from the pod to service, service
to service, and workload to service dashboards and creates a new
dashboard for workload to workload metrics.

Signed-off-by: jaellio <jaellio@microsoft.com>