feat(cli): add uninstall cluster-wide-resources cmd
This adds a command `osm uninstall cluster-wide-resources`,
which uninstalls cluster-wide resources that may have been
left over after osm was previously uninstalled
(through `osm uninstall mesh`, through helm, or through
uninstallation in a managed environment).

When `osm uninstall mesh` is run, the helm post-delete hook
is invoked as part of the uninstall process. This deletes:
(1) osm control plane pods (in the osm namespace)
(2) osm control plane secrets (in the osm namespace)
(3) mutating webhook configurations
(4) validating webhook configurations
(5) osm control plane mesh config

OSM CRDs and SMI CRDs are not deleted by the helm
post-delete hook.

Additionally, all of the above are not deleted when osm
is uninstalled from a cluster in a managed environment,
as the helm post-delete hook is typically not executed
in such environments.

This new subcommand ensures that the following are deleted
from the cluster, whether the cluster is managed or not:
(1) OSM CRDs
(2) SMI CRDs
(3) OSM MutatingWebhookConfigurations
(4) OSM ValidatingWebhookConfigurations
(5) OSM Secrets

Resolves #4255.

Signed-off-by: Johnson Shi <Johnson.Shi@microsoft.com>